persistence/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyYAML==6.0.1
GitPython==3.1.40
structlog==24.1.0
Werkzeug==3.0.1